Broke up PrePublish method that was too complex
